requirements.txt
matplotlib>=1.5,<1.6
requests>=2.10,<2.11
progressbar2>=3.6,<3.7
